requirements.txt
annotated-types==0.6.0
boto3==1.34.68
botocore==1.34.68
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Deprecated==1.2.14
google-auth==2.29.0
idna==3.6
jmespath==1.0.1
kubernetes==29.0.0
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pydash==8.0.0
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
typing_extensions==4.9.0
urllib3==2.2.1
websocket-client==1.7.0
wrapt==1.16.0
PyYAML
pydash
PyGithub
python-dotenv
azure-identity
kubernetes